git-svn-id: https://svn.kuali.org/repos/student/deploymentlab/branches/sts-2.5.1@17605 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-core/ks-core-ui/src/main/java/org/kuali/student/core/organization/assembly/OrgPersonRelationAssembler.java
@@ -36,6 +36,7 @@ import org.kuali.student.core.assembly.data.Metadata;
 import org.kuali.student.core.assembly.data.QueryPath;
 import org.kuali.student.core.assembly.data.SaveResult;
 import org.kuali.student.core.assembly.data.Data.Property;
+import org.kuali.student.core.assembly.util.AssemblerUtils;
 import org.kuali.student.core.dto.MetaInfo;
 import org.kuali.student.core.exceptions.DoesNotExistException;
 import org.kuali.student.core.organization.assembly.data.server.org.OrgHelper;
@@ -133,9 +134,10 @@ public class OrgPersonRelationAssembler implements Assembler<Data, OrgPersonHelp
                     } catch (Exception e) {
                         throw new AssemblyException();
                     }
+                    AssemblerUtils.setUpdated(orgPersonHelper.getData(), false);
                 }
             }
-            else if(isDeleted(orgPersonHelper.getData())){
+            else if(isDeleted(orgPersonHelper.getData())&&orgPersonHelper.getId()!=null){
                 try{
                     orgService.removeOrgPersonRelation(orgPersonHelper.getId());
                     propertyIter.remove();
@@ -157,6 +159,7 @@ public class OrgPersonRelationAssembler implements Assembler<Data, OrgPersonHelp
                 	LOG.error(e);
                     throw new AssemblyException();
                 }
+                AssemblerUtils.setCreated(orgPersonHelper.getData(), false);
             }
